MethylationBetaValue_fix
The gene symbol is chosen from the methylated site. If the site's
coordinates are in the genomic region of the considered gene, the gene
name enhances the gene_symbol field, otherwise the gene, whose
coordinates are closest to the methylated site, is chosen. The
entrez_id and the strand are also extracted from the chosen gene.